pyproject.toml
[tool.poetry]
name = "chip_logistics"
version = "0.1.0"
description = "Telegram-bot for logistic prices calculations for WelcomeToRussia company."
authors = ["butvinm <butvin.mihail@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
pydantic = "^2.3.0"
aiohttp = "^3.8.5"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
aiogram = ">=3"
aiogram_deta = { git = "https://github.com/butvinm/aiogram_deta.git", branch = "flexible-deps"}

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.18.0"
mypy = "^1.5.1"
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
